+++ ClientApp/src/App.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import { Route } from 'react-router';
 import { Layout } from './components/Layout';
 import HomePage from './components/home/Page';
-import LeaguesPage from './components/leagues/Page';
+import SeasonsPage from './components/seasons/Page';
 
 export default class App extends Component {
   displayName = App.name
@@ -11,7 +11,7 @@ export default class App extends Component {
     return (
       <Layout>
         <Route exact path='/' component={HomePage} />
-        <Route exact path='/leagues' component={LeaguesPage} />
+        <Route exact path='/seasons' component={SeasonsPage} />
       </Layout>
     );
   }
